+++ packages/vite-plugin/tests/e2e/mv3-vite-hmr-background-define/src/background.ts
@@ -1,0 +1,11 @@
+// This fixes `self`'s type.
+declare const self: ServiceWorkerGlobalScope
+declare const __HELLO_WORLD__: string
+
+export {}
+
+console.log('defined', __HELLO_WORLD__)
+
+chrome.runtime.onInstalled.addListener(() => {
+  chrome.runtime.openOptionsPage()
+})