+++ packages/vite-plugin/tests/e2e/mv3-vite-hmr-background-define/vite.config.ts
@@ -1,0 +1,11 @@
+import { crx } from 'src/index'
+import { defineConfig } from 'vite'
+import manifest from './manifest.json'
+
+export default defineConfig({
+  build: { minify: false },
+  clearScreen: false,
+  logLevel: 'error',
+  plugins: [crx({ manifest })],
+  define: { __HELLO_WORLD__: JSON.stringify('Hello, world!') },
+})